Updated average_error.py
Updated the function to calculate the difference between the simulated cases and the synthetic data and display the differences each time.

Added documentation for each function to help the readers gain a better understanding of each function. 

Further work would be to calculate an average for the differences between the errors to help measure the errors in the measurement of the number of infections during a pandemic.